Move icon/vertical_diffusion/ to vertical_diffusion/icon/
Convert relative import to absolute, update icon_physics.py and parameters.py.
