+++ jcm/physics/icon/parameters.py
@@ -12,7 +12,7 @@ from jcm.physics.convection.icon import ConvectionParameters
 from jcm.physics.clouds.icon import CloudParameters, MicrophysicsParameters
 from jcm.physics.gravity_waves.icon import GravityWaveParameters
 from .radiation import RadiationParameters
-from .vertical_diffusion.vertical_diffusion_types import VDiffParameters
+from jcm.physics.vertical_diffusion.icon.vertical_diffusion_types import VDiffParameters
 from jcm.physics.surface.icon import SurfaceParameters
 from jcm.physics.aerosol.icon.aerosol_params import AerosolParameters
